Update sudokumain.c
Renamed "nosolution" to "no_solution"
Added several probably unnecessary checks in a failed attempt at debugging the code 
Added a section to reset the grid to a Zero Matrix if no solution is found, however the code still hangs before reaching that point
Gave use to the pointer "*p". Guess I forgot to do that when originally tackling this code.
Rebuilt the Message Loop. Not sure why, that part I did a while ago and I don't remember the purpose

Notes to self: Hardcoding the "valid" int to 0 or 1 produces the "No Solution" and "Success" results, respectively. However, setting that to the isValid function makes the program commit suicide still when there is no solution. I've provided blocks to prevent the isValid from trying to process array elements that aren't on the array, but that doesn't fix it